removed dirty file logic from main_window
