+++ qiskit_metal/_gui/widgets/developer_mode/developer_mode.py
@@ -1,5 +1,4 @@
 
-from PySide2.QtCore import QTimer, Qt
 from PySide2.QtWidgets import (QAction, QTreeView)
 from typing import Callable
 class DeveloperMode():
@@ -12,6 +11,7 @@ class DeveloperMode():
 
 
     def setup_developer_mode(self, rebuild_action: QAction, rebuild_function: Callable, qlibrary_tree_view: QTreeView):
+        pass
         # add delegate to tree
         # hook up RebuildACtion's trigger to clean all files
         #     source_model.file_dirtied_signal.connect(tree.update)
@@ -22,7 +22,8 @@ class DeveloperMode():
         # hook up tree make QAction Icon red if dirty
 
 
-    def clean_and_remove_developer_mode:
+    def clean_and_remove_developer_mode(self):
+        pass
         # set empty delegate to tree
             #    undo      source_model.file_dirtied_signal.connect(tree.update)
 # source_model.file_cleaned_signal.connect(tree.update)
@@ -31,3 +32,34 @@ class DeveloperMode():
         # return original icon to QAction
         # replace rebuild singla/slot with oiriginal rebuild func
         # unhookup tree to file action
+
+
+    def _refresh_component_build(self, qis_abs_path):
+        """Reresh build for a component along a given path.
+
+        Args:
+            qis_abs_path (str): Absolute component path.
+        """
+        self.design.reload_and_rebuild_components(qis_abs_path)
+        # Table models
+        self.ui.tableComponents.model().refresh()
+
+        # Redraw plots
+        self.refresh_plot()
+        self.autoscale()
+
+
+
+    def refresh_everything(self):
+        """Refresh everything."""
+
+        df = self.ui.dockLibrary.library_model.dirtied_files
+        values = {list(df[k])[0] for k in df.keys()}
+
+        for file in values:  # dirtied_files size changes during clean_file
+            if '.py' in file:
+                file = file[file.index('qiskit_metal'):]
+                self.design.reload_and_rebuild_components(file)
+                self.ui.dockLibrary.library_model.clean_file(file)
+        self.refresh()
+        self.autoscale()
